Moved some packages and fixed some imports to get build_master_images_table.py running under the new package structure. No optimizatioons yet. For Ticket # 103.
